+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fix default i18n option for Modal dialogs
 - Fix Slideshow initially shows last slide first
 - Fix Height Match component does not change height while updating
+- Fix Slider reactivity when a slide is hidden/shown
 
 ## 3.16.3 (February 27, 2023)